[wip] add branch for resolving dependencies of a library
