+++ cabal-audit.cabal
@@ -26,9 +26,8 @@ tested-with:        GHC ==9.4.8 || ==9.6.4 || ==9.8.2
 
 common common-all
   ghc-options:
-    -Wall -Wcompat -Widentities -Wincomplete-record-updates
-    -Wincomplete-uni-patterns -Wredundant-constraints
-    -fmax-relevant-binds=0 -fno-show-valid-hole-fits
+    -Wall -Wcompat -Wunused-packages -fmax-relevant-binds=0
+    -fno-show-valid-hole-fits
 
   if impl(ghc >=9.6.1)
     ghc-options: -fno-show-error-context
@@ -40,9 +39,6 @@ common common-all
     DerivingStrategies
     DerivingVia
     DuplicateRecordFields
-    EmptyCase
-    FlexibleContexts
-    FlexibleInstances
     GADTs
     ImportQualifiedPost
     LambdaCase
@@ -68,6 +64,7 @@ library
   exposed-modules:
     Control.Effect.Pretty
     Distribution.Audit
+    Distribution.Audit.Parser
     Security.Advisories.Cabal
     Security.Advisories.SBom.CycloneDX
     Security.Advisories.SBom.Types
@@ -78,19 +75,16 @@ library
     , bytestring
     , Cabal
     , cabal-install
+    , cabal-install-solver
     , chronos
     , colourista
     , containers
-    , filepath
     , fused-effects
-    , hsec-core
     , hsec-tools
-    , http-client
     , kan-extensions
     , optparse-applicative
     , pretty
     , process
-    , temporary
     , text
     , transformers
     , unliftio
@@ -99,7 +93,7 @@ library
     , vector
 
   hs-source-dirs:   src
-  default-language: Haskell2010
+  default-language: GHC2021
 
 executable cabal-audit
   import:           common-all
